Rework project structure

+++ src/main/java/com/sams/demo/repository/QuestionRepository.java
@@ -1,13 +1,13 @@
 package com.sams.demo.repository;
 
-import com.sams.demo.entity.Question;
+import com.sams.demo.model.entity.Question;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuestionRepository extends JpaRepository<Question, Integer> {
+public interface QuestionRepository extends JpaRepository<Question, Long> {
 
     Page<Question> findAll(Pageable pageable);
 }
